modified the all js files as requested by the reviewer
