+++ package.json
@@ -25,6 +25,7 @@
     "hint": "^7.1.10",
     "html-webpack-plugin": "^5.5.3",
     "jest": "^29.6.4",
+    "jsdom": "^22.1.0",
     "style-loader": "^3.3.3",
     "stylelint": "^13.13.1",
     "stylelint-config-standard": "^21.0.0",
