+++ src/index.js
@@ -1,6 +1,6 @@
 import './style.css';
 import './assets/styletwo.css';
-import homeContent from './module/homepagedisplayer.js';
+import homeContent  from './module/homepagedisplayer.js';
 import { getLike } from './module/likeshandler.js';
 
 window.addEventListener('load', async () => {
